feat(client): render exclusively on DirectX 12
Flip the renderer to DirectX 12 only: the state manager translates every
draw to the DX12 backend, the shader pool replaces the fixed-function
pipeline, and the vendored Direct3D 9 SDK and all fallback paths are gone.
Fixes the render regressions surfaced during in-game validation (fonts,
UI edges, granny armor, minimap, dungeon lighting, tree leaves, blend and
fade passes, white materials) and the frame-resource and fence defects
found by the backend audit.

+++ Lead-Client-Source/EterLib/GrpDescriptorRingDX12.h
@@ -36,7 +36,7 @@ class CGraphicDescriptorRingDX12
 	private:
 		struct TFrameSpan
 		{
-			UINT	uHead;
+			UINT64	uHead;
 			UINT64	uFenceValue;
 		};
 
@@ -45,8 +45,8 @@ class CGraphicDescriptorRingDX12
 		ID3D12DescriptorHeap*	m_pkHeap;
 		UINT					m_uCapacity;
 		UINT					m_uIncrementSize;
-		UINT					m_uHead;
-		UINT					m_uTail;
+		UINT64					m_uHead;
+		UINT64					m_uTail;
 		TFrameSpan				m_akSpans[MAX_SPANS];
 		UINT					m_uSpanCount;
 };
